Split up project into modules
