+++ test_birdyboard.py
@@ -11,28 +11,11 @@ class TestBirdyBoard(unittest.TestCase):
       # self.delete_message = birdy.delete_chirp(self.message)
       # self.read_message = birdy.read_chirp()
 
-    def test_new_user_is_a_csv_list(self):
-      self.birdy.create_user("Juniper Jones", "birdistheword")
-      user_data = list()
-      users = self.birdy.read_users()
-      user_data.extend(users)
-      self.assertIn("Juniper Jones", user_data)
 
-    def test_chirp_is_written_to_chirps_file(self):
-      self.write_message
-      self.assertIn(self.message,'chirps.csv')
-      self.delete_message
 
-    def test_chirp_is_removed_from_chirps_file(self):
-      self.write_message
-      self.delete_message
-      self.assertNotIn(self.message, 'chirps.csv')
 
-    def test_view_chirps_returns_chirps(self):
-      self.write_message
-      message_list = birdy.read_chirps()
-      self.delete_message
-      self.assertTrue(message_list == [str])
+
+
 
     # def test_private_chirp_is_private(self):
     #   pass
